# back-end/src/main/resources/api-docs/openapi3.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "당구사랑",
    "description" : "당구사랑 API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:9000"
  } ],
...
  "paths" : {
...
    "/api/v1/contest/{id}/join" : {
      "post" : {
        "tags" : [ "api" ],
        "operationId" : "contest-joinerror-already-joinerror-cant-join",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "대회 아이디",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "JWT 토큰",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI1IiwiaWF0IjoxNjMxMTM2MDA4LCJleHAiOjE2MzExMzc4MDh9.kXWONACN7YsORX5c71GnJ0G3wr4p6iUFpAcR74YHBjo"
        } ],
        "requestBody" : {
          "content" : {
            "application/json;charset=UTF-8" : {
              "schema" : {
                "$ref" : "#/components/schemas/api-v1-contest-id-join-405702609"
              },
              "examples" : {
                "contest-join" : {
                  "value" : "{\r\n  \"handicap\" : 30\r\n}"
                },
                "error-already-join" : {
                  "value" : "{\r\n  \"handicap\" : 30\r\n}"
                },
                "error-cant-join" : {
                  "value" : "{\r\n  \"handicap\" : 30\r\n}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "200"
          },
          "400" : {
            "description" : "400",
            "content" : {
              "application/json;charset=UTF-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/api-v1-contest-id-join-1153901879"
                },
                "examples" : {
                  "error-already-join" : {
                    "value" : "{\r\n  \"code\" : 400,\r\n  \"message\" : \"이미 참가한 대회입니다.\"\r\n}"
                  },
                  "error-cant-join" : {
                    "value" : "{\r\n  \"code\" : 400,\r\n  \"message\" : \"대회에 참가할 수 없는 상태입니다.\"\r\n}"
                  }
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "api" ],
        "operationId" : "contest-join-cancel",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "대회 아이디",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "JWT 토큰",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI0IiwiaWF0IjoxNjMxMTM2MDA4LCJleHAiOjE2MzExMzc4MDh9.MH2vRifbYzPqR6iHzcMGXlIjPalOUfP_kKXt-gUNpXs"
        } ],
        "responses" : {
          "200" : {
            "description" : "200"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "api-v1-contest-id-join-1153901879" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "number",
            "description" : "에러 코드"
          },
          "message" : {
            "type" : "string",
            "description" : "에러 메세지"
          }
        }
      },
...
      "api-v1-contest-id-join-405702609" : {
        "type" : "object",
        "properties" : {
          "handicap" : {
            "type" : "number",
            "description" : "참가 핸디캡"
          }
        }
      },
...
    }
  }
}